add method

+++ core/rawdb/accessors_chain.go
@@ -163,9 +163,9 @@ func WriteHeaderNumber(db ethdb.KeyValueWriter, hash common.Hash, number uint64)
 }
 
 // WriteBlockStateByNumber store the hash -> state 
-func WriteBlockStateByNumber(db ethdb.KeyValueWriter,hash common.Hash,state int) error {
+func WriteBlockStateByNumber(db ethdb.KeyValueWriter,hash common.Hash,state uint64) error {
 	key := headerNumberStateKey(hash)
-	data := encodeBlockNumber(uint64(state))
+	data := encodeBlockNumber(state)
 	if err := db.Put(key, data); err != nil {
 		return err
 		log.Crit("Failed to store hash to number mapping", "err", err)
